+++ ci/tox/local-run.sh
@@ -8,10 +8,6 @@ function check_var() {
   grep -q "^$1=" .env || { echo "Var $1 missing both from environment and .env file"; return 1; }
 }
 
-function build_image() {
-  docker-compose build --build-arg user_id="$(id -u)" --build-arg group_id="$(id -g)" tox
-}
-
 cd "$(git rev-parse --show-toplevel)"/ci/tox/
 check_var GIT_VERSION
 check_var PYTHON_VERSION
@@ -21,10 +17,9 @@ set -x
 hash=$(git rev-parse HEAD:ci/tox)
 if git diff-index --quiet HEAD .; then
   export DIRECTORY_HASH="$hash"
-  docker-compose pull tox || build_image
 else
   export DIRECTORY_HASH="$hash"-dirty
-  build_image
 fi
 
+docker-compose build --build-arg user_id="$(id -u)" --build-arg group_id="$(id -g)" tox
 docker-compose up --exit-code-from=tox tox